requirements.txt
asgiref==3.7.2
Django==5.0.2
pillow==10.2.0
setuptools==69.1.1
sqlparse==0.4.4
whitenoise==6.9.0
django-ckeditor==6.7.2
django-cloudinary-storage==0.3.0
dj-database-url==2.3.0
cloudinary==1.42.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1